Rename fpx-auth lib module to fp-services and add a utility for making requests
